+++ app/src/main/java/org/b3log/siyuan/activity/RedirectUriReceiverActivity.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package org.b3log.siyuan;
+package org.b3log.siyuan.activity;
 
 import android.content.Intent;
 import android.net.Uri;
