+++ app/src/main/java/org/b3log/siyuan/config/AppConfig.java
@@ -15,7 +15,9 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package org.b3log.siyuan;
+package org.b3log.siyuan.config;
+
+import org.b3log.siyuan.util.Utils;
 
 /**
  * Application configuration constants.
