+++ app/src/main/java/org/b3log/siyuan/manager/AppAssetManager.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package org.b3log.siyuan;
+package org.b3log.siyuan.manager;
 
 import android.content.Context;
 import android.util.Log;
@@ -24,6 +24,7 @@ import com.blankj.utilcode.util.StringUtils;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
+import org.b3log.siyuan.util.Utils;
 
 import java.io.File;
 import java.nio.charset.StandardCharsets;
