+++ app/src/main/java/org/b3log/siyuan/manager/KernelManager.java
@@ -15,12 +15,13 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package org.b3log.siyuan;
+package org.b3log.siyuan.manager;
 
 import android.content.Context;
-import android.content.pm.PackageManager;
 import android.os.Build;
 import android.util.Log;
+
+import org.b3log.siyuan.util.Utils;
 
 import java.util.TimeZone;
