+++ app/src/main/java/org/b3log/siyuan/util/AndroidBug5497Workaround.java
@@ -1,4 +1,4 @@
-package org.b3log.siyuan;
+package org.b3log.siyuan.util;
 
 import android.annotation.SuppressLint;
 import android.app.Activity;
