+++ app/src/main/java/org/b3log/siyuan/util/Utils.java
@@ -15,7 +15,9 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package org.b3log.siyuan;
+package org.b3log.siyuan.util;
+
+import org.b3log.siyuan.BuildConfig;
 
 import android.app.Activity;
 import android.content.Context;
